+++ src/aws/cloud-formation-wrapper.ts
@@ -15,33 +15,37 @@ class CloudFormationWrapper {
     }
 
     /**
-     * Gets rest API id from CloudFormation stack
+     * Gets rest API id from CloudFormation stack or nested stack
      */
     public async getApiId(domain: DomainConfig, stackName: string): Promise<string> {
-        let LogicalResourceId = "ApiGatewayRestApi";
+        let logicalResourceId = "ApiGatewayRestApi";
         if (domain.apiType === Globals.apiTypes.http) {
-            LogicalResourceId = "HttpApi";
+            logicalResourceId = "HttpApi";
         }
         if (domain.apiType === Globals.apiTypes.websocket) {
-            LogicalResourceId = "WebsocketsApi";
+            logicalResourceId = "WebsocketsApi";
         }
-
-        const params = {
-            LogicalResourceId,
-            StackName: stackName,
-        };
 
         let response;
         try {
-            response = await throttledCall(this.provider, "describeStackResource", params);
-        } catch (err) {
-            throw new Error(`Failed to find CloudFormation resources with an error: ${err}\n`);
+            // trying to get information for specified stack name
+            response = await this.getStack(logicalResourceId, stackName);
+        } catch {
+            // in case error trying to get information from the some of nested stacks
+            response = await this.getNestedStack(logicalResourceId, stackName);
+        }
+
+        if (!response) {
+            throw new Error(`Failed to find a stack ${stackName}\n`);
         }
 
         const apiId = response.StackResourceDetail.PhysicalResourceId;
         if (!apiId) {
             throw new Error(`No ApiId associated with CloudFormation stack ${stackName}`);
         }
+
+        Globals.logInfo(`Found apiId: ${apiId}`, domain.givenDomainName, false);
+
         return apiId;
     }
 
@@ -63,6 +67,51 @@ class CloudFormationWrapper {
         // converting a list of unique values to dict
         // [{Name: "export-name", Value: "export-value"}, ...] - > {"export-name": "export-value"}
         return filteredExports.reduce((prev, current) => ({...prev, [current.Name]: current.Value}), {});
+    }
+
+    /**
+     * Returns a description of the specified resource in the specified stack.
+     */
+    public async getStack(logicalResourceId: string, stackName: string) {
+        try {
+            return await throttledCall(this.provider, "describeStackResource", {
+                LogicalResourceId: logicalResourceId,
+                StackName: stackName,
+            });
+        } catch (err) {
+            throw new Error(`Failed to find CloudFormation resources with an error: ${err}\n`);
+        }
+    }
+
+    /**
+     * Returns a description of the specified resource in the specified nested stack.
+     */
+    public async getNestedStack(logicalResourceId: string, stackName?: string) {
+        // get all stacks from the CloudFormation
+        const stacks = await getAWSPagedResults(
+            this.provider,
+            "describeStacks",
+            "Stacks",
+            "NextToken",
+            "NextToken",
+            {},
+        );
+
+        // filter stacks by given stackName
+        const filteredStackNames = stacks
+            .map((stack) => stack.StackName) // collect list of stack names
+            .filter((name) => name.includes(stackName)); // filter by stackName
+
+        let response;
+        for (const name of filteredStackNames) {
+            try {
+                response = await this.getStack(logicalResourceId, name);
+                break;
+            } catch (err) {
+                Globals.logError(err);
+            }
+        }
+        return response;
     }
 }